Crate dhstore

DHStore: A personal content management system.

Re-exports

pub use common::ID;

pub use errors::Error;

Modules

errors
Error definitions

hash
Structures and functions related to hashing.

log
Log utilities.

Structs

FileBlobStorage
Filesystem-based blob storage implementation.

MemoryIndex
The in-memory index, that loads all objects from the disk on startup.

Object
A schema object, i.e. either a dictionary or a list of properties.

Store
Main structure, representing the whole system.

Enums

ObjectData
The types of object known to the index.

Property
Values that appear in an object's metadata.

Traits

BlobStorage
Trait for the blob storage backends, that handle the specifics of storing blobs. A blob is an unnamed sequence of bytes, which constitute parts of some file's contents.

EnumerableBlobStorage
Additional trait for a BlobStorage that knows how to enumerate all the blobs it has.

ObjectIndex
Trait for the index of schema objects.

Functions

create
Creates a new store on disk.

open
Opens a directory.

permanode

Type Definitions

Dict

List