Struct dhstore::Store

pub struct Store<S: BlobStorage, I: ObjectIndex> { /* fields omitted */ }

Main structure, representing the whole system.

Methods

impl<S: BlobStorage, I: ObjectIndex> Store<S, I>

pub fn new(storage: S, index: I) -> Store<S, I>

Creates a store from a given blob storage and object index.

pub fn add_blob<R: Read>(&mut self, reader: R) -> Result<ID>

Low-level; adds a blob to the blob storage.

To cut a blob into chunks, add them to the blob storage, and return a list object of them, use Store::add_file().

pub fn get_blob(&self, id: &ID) -> Result<Option<Box<[u8]>>>

Low-level; gets a single blob from the blob storage.

pub fn get_object(&self, id: &ID) -> Result<Option<&Object>>

Low-level; gets a single object from the index by its ID.

pub fn add_file<R: Read>(&mut self, reader: R) -> Result<(ID, usize)>

Cuts a file into chunks and add a list object of them to the index.

pub fn add<P: AsRef<Path>>(&mut self, path: P) -> Result<ID>

Adds a file or directory recursively, representing directories as dicts and files as lists of blobs.

pub fn verify(&mut self) -> Result<()>

Checks the blobs and objects for errors.

pub fn print_object(&self, id: &ID, limit: Option<usize>) -> Result<()>

Pretty-prints objects recursively.

If limit is not None, it is the maximum depth of nested objects we'll print; for example, Some(1) means that objects directly referenced from the given one will be expanded, but not objects referenced from those.

impl<S: EnumerableBlobStorage, I: ObjectIndex> Store<S, I>

pub fn collect_garbage(&mut self) -> Result<()>