Trait dhstore::ObjectIndex

pub trait ObjectIndex {
    fn add(&mut self, data: ObjectData) -> Result<ID>;

    fn get_object(&self, id: &ID) -> Result<Option<&Object>>;

    fn verify(&mut self) -> Result<()>;

    fn collect_garbage(&mut self) -> Result<HashSet<ID>>;
}

Trait for the index of schema objects.

This is a sort of database that can store Objects and knows how to make sense of them and query them efficiently.

Required Methods

fn add(&mut self, data: ObjectData) -> Result<ID>

Hashes an object and adds it to the index.

fn get_object(&self, id: &ID) -> Result<Option<&Object>>

Gets an object from its hash.

fn verify(&mut self) -> Result<()>

Checks the index for errors.

fn collect_garbage(&mut self) -> Result<HashSet<ID>>

Deletes unreferenced objects and returns the set of blobs to keep.